Configurable `tagName` value
It would be nice to be able to set the name of the struct tag used when parsing the column name from the struct. This pull request does this in the most simple way I could think of -- by changing the `const tagName` to `var TagName`.

I'm working on a "Resource" struct that provides the bridge between my database rows and my API response -- and I've already got a couple of tags on my struct fields, one called `json`, for `endcoding/json`, and one called `db` for jmoiron/sqlx. Ideally, I'd like to re-use this `db` tag as they follow the same rules as your library, which I am only using for the handy `Columns()` function.

Thank you for considering this humble pull request.